# fix: Replace broken generateKey with fast-json-stable-stringify for pseudo props