--- a/.changeset/kind-mice-shake.md
+++ b/.changeset/kind-mice-shake.md
@@ -1,0 +1,6 @@
+---
+"@kuma-ui/compiler": patch
+"@kuma-ui/core": patch
+---
+
+fix: Replace broken generateKey with fast-json-stable-stringify for pseudo props