Closes #246 

## Problem

In short, Kuma UI's pseudo props support is basically broken.

Before this PR,  `generateKey` was generating something like `_hover:[object Object]|p:2`. After this PR, it uses `fast-json-stable-stringify` and it generates `{"_hover":{"color":"red"},"p":2}`, which includes nested values.
This is critical when using pseudo props. Because of this, when using pseudo props, most of the time styling was broken since actual values of pseudo stlying is not taken into account when taking a hash for style caching.

I used the following example to confirm the bug:
```tsx
import { Input } from "@kuma-ui/core";

function Demo() {
  return (
    <>
      <Input m={2} _valid={{ bgColor: 'red' }} />
      <Input m={2} _valid={{ bgColor: 'blue' }} />
    </>
  );
}
```
### Before this PR
<img width="320" alt="before" src="https://github.com/kuma-ui/kuma-ui/assets/2931577/7fd99783-8625-4180-b2fa-869e8fcda54d">

Both input elements are red and "bgColor: blue" is completely ignored



### After this PR
<img width="343" alt="after" src="https://github.com/kuma-ui/kuma-ui/assets/2931577/49c957aa-5663-4c7b-89df-542e5baaa860">

Colors are expected

## Solution

Use `fast-json-stable-stringify` instead of using our own implementation of generateKey.

Note that this logic can be hot execution path when Kuma UI fails to extract static styles.
For example, emotion is tying to optimize [the whole "serialize styles" thing](https://github.com/emotion-js/emotion/blob/f3b268f7c52103979402da919c9c0dd3f9e0e189/packages/serialize/src/index.js#L310-L388) and in my experience of using Emotion in large scale website, this was a heavy bottleneck. We might want to revisit this choice in the future but we should fix a bug first.

## The choice of fast-json-stable-stringify

I ran [this benchmark](https://github.com/epoberezkin/fast-json-stable-stringify#benchmark) in my MacBook:
```
fast-json-stable-stringify x 61,432 ops/sec ±0.20% (101 runs sampled)
json-stable-stringify x 48,248 ops/sec ±0.12% (97 runs sampled)
fast-stable-stringify x 65,665 ops/sec ±0.10% (97 runs sampled)
faster-stable-stringify x 55,941 ops/sec ±1.77% (96 runs sampled)
The fastest is fast-stable-stringify
```

For now I would choose "fast-json-stable-stringify" for since I found it's used in [`@jest/transform`](https://github.com/facebook/jest/blob/0fd5b1c37555f485c56a6ad2d6b010a72204f9f6/packages/jest-transform/src/ScriptTransformer.ts#L14-L14).
Also note that this relies on the fact **Kuma UI never relies on prop ordering even for nested cases**.