@naruaway 2023-08-08
Discussed with the core maintainers, we decided to vendor fast-json-stable-stringify in the codebase to avoid adding new package dependency.
Since we need it for two different packages, for now we duplicate the implementation with code comment. In the future we should come up with a way to share intenal utils across packages without exposing them to end users